% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coord_golden}
\alias{coord_golden}
\title{Set aspect ratio according to the golden ratio}
\usage{
coord_golden(xvals, yvals, orientation = "horizontal")
}
\arguments{
\item{xvals}{Vector containing the values plotted along the X axis}

\item{yvals}{Vector containing the values plotted along the Y axis}

\item{orientation}{Whether the golden ratio should be "horizontal" or
"vertical" (default: horizontal)}
}
\description{
\code{coord_golden} sets the aspect ratio of the axes in the plot according
to the golden ratio.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x=hp, y=mpg)) + geom_point() +
         coord_golden(xvals=mtcars$hp, yvals=mtcars$mpg)
p
}
\seealso{
ratio_golden
}

