% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_remove_background.R
\name{gg_remove_background}
\alias{gg_remove_background}
\alias{gg_remove_background_canvas}
\alias{gg_remove_background_facet}
\alias{gg_remove_background_legend}
\alias{gg_remove_background_panel}
\title{Remove a plot's background.}
\usage{
gg_remove_background(plot = last_plot(), canvas = TRUE, panel = FALSE,
  legend = FALSE, facet = FALSE)

gg_remove_background_canvas(plot = last_plot())

gg_remove_background_panel(plot = last_plot())

gg_remove_background_legend(plot = last_plot())

gg_remove_background_facet(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{canvas}{Whether or not to remove the background from the entire image (default: \code{TRUE})}

\item{panel}{Whether or not to remove the background from the panel (i.e., the plotting area; default: \code{FALSE})}

\item{legend}{Whether or not to remove the legend's background
(default: \code{FALSE})}

\item{facet}{Whether or not to remove the background from facet labels
(default: \code{FALSE})}
}
\value{
A modified ggplot2 plot
}
\description{
\code{gg_remove_background} removes the background from the given plot. This
is useful for creating plots with transparent backgrounds

\code{gg_remove_background_canvas} Removes TODO

\code{gg_remove_background_panel} Removes TODO

\code{gg_remove_background_legend} Removes TODO

\code{gg_remove_background_facet} Removes TODO
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1)
gg_remove_background(plot = p)

}

