% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_remove_grid.R
\name{gg_remove_grid}
\alias{gg_remove_grid}
\alias{gg_remove_grid_major}
\alias{gg_remove_grid_minor}
\alias{gg_remove_grid_x}
\alias{gg_remove_grid_y}
\title{Remove grid lines from a plot}
\usage{
gg_remove_grid(plot = last_plot(), major.x = TRUE, major.y = TRUE,
  minor.x = TRUE, minor.y = TRUE)

gg_remove_grid_major(plot = last_plot(), major.x = TRUE, major.y = TRUE)

gg_remove_grid_minor(plot = last_plot(), minor.x = TRUE, minor.y = TRUE)

gg_remove_grid_x(plot = last_plot(), major = TRUE, minor = TRUE)

gg_remove_grid_y(plot = last_plot(), major = TRUE, minor = TRUE)
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{major.x}{Whether or not to remove grid lines at major breakpoints along the X axis (default: \code{TRUE})}

\item{major.y}{Whether or not to remove grid lines at major breakpoints along the Y axis (default: \code{TRUE})}

\item{minor.x}{Whether or not to remove grid lines at minor breakpoints along the X axis (default: \code{TRUE})}

\item{minor.y}{Whether or not to remove grid lines at minor breakpoints along the Y axis (default: \code{TRUE})}
}
\value{
A modified ggplot2 plot
}
\description{
\code{gg_remove_grid} removes either the major or minor grid lines on either
axis.

\code{gg_remove_grid_major} Removes grid lines at major breaks

\code{gg_remove_grid_minor} Removes grid lines at minor breaks

\code{gg_remove_grid_x} Removes grid lines at major and minor
breaks on the X axis

\code{gg_remove_grid_y} Removes grid lines at major and minor
breaks on the X axis
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1)
gg_remove_grid(plot = p)

}

