% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_remove_ticks.R
\name{gg_remove_ticks}
\alias{gg_remove_ticks}
\alias{gg_remove_ticks_x}
\alias{gg_remove_ticks_y}
\title{Remove tick marks from a plot's axes}
\usage{
gg_remove_ticks(plot = last_plot(), x = TRUE, y = TRUE)

gg_remove_ticks_x(plot = last_plot())

gg_remove_ticks_y(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{x}{Whether or not to remove X-axis ticks (default: \code{TRUE})}

\item{y}{Whether or not to remove Y-axis ticks (default: \code{TRUE})}
}
\value{
A modified ggplot2 plot
}
\description{
\code{gg_remove_ticks} can remove ticks on either the X or Y axes. By
default, all tick marks are removed.

\code{gg_remove_ticks_x} removes tick marks from the X axis

\code{gg_remove_ticks_y} removes tick marks from the Y axis
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1)
gg_remove_ticks(plot = p)

}

