% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_remove_titles.R
\name{gg_remove_titles}
\alias{gg_remove_caption}
\alias{gg_remove_subtitle}
\alias{gg_remove_title}
\alias{gg_remove_title.x}
\alias{gg_remove_title.y}
\alias{gg_remove_titles}
\title{Remove titles from a plot}
\usage{
gg_remove_titles(plot = last_plot(), title = TRUE, subtitle = TRUE,
  caption = TRUE, x = TRUE, y = TRUE)

gg_remove_title(plot = last_plot())

gg_remove_subtitle(plot = last_plot())

gg_remove_caption(plot = last_plot())

gg_remove_title.x(plot = last_plot())

gg_remove_title.y(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{title}{Whether or not to remove the main title (default: \code{TRUE})}

\item{subtitle}{Whether or not to remove the subtitle (default: \code{TRUE})}

\item{caption}{Whether or not to remove the caption (default: \code{TRUE})}

\item{x}{Whether or not to remove the X axis title (default: \code{TRUE})}

\item{y}{Whether or not to remove the Y axis title (default: \code{TRUE})}
}
\value{
A modified ggplot2 plot
}
\description{
\code{gg_remove_titles} removes title elements such as the main title,
subtitle, caption, or axis titles. By default, all are removed.

\code{gg_remove_title} Removes the plot's main title

\code{gg_remove_subtitle} Removes the plot's subtitle

\code{gg_remove_caption} Removes the plot caption

\code{gg_remove_title.x} Removes the X-axis title

\code{gg_remove_title.y} Removes the Y-axis title
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1) +
    labs(title = "Fuel Efficiency by Engine Type",
         subtitle = "Larger engines are less efficient",
         caption = "Source: Motor Trend",
         x = "Number of Cylinders",
         y = "Fuel Efficiency (MPG)")
gg_remove_titles(plot = p)

}

