% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_rescale.R
\name{gg_rescale}
\alias{gg_rescale}
\alias{gg_rescale_golden}
\alias{gg_rescale_square}
\title{Rescale a ggplot2 plot so that axes follow a given ratio}
\usage{
gg_rescale(plot = last_plot(), ratio)

gg_rescale_golden(plot = last_plot(), orient = "h")

gg_rescale_square(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{ratio}{The aspect ratio to use for the axes. This is independent of
units used in the plot, so the size of the X axis will be ratio times the
total size of the Y axis.}

\item{orient}{Whether the golden ratio should be horizontal
(\code{h}; default) or vertical (\code{v})}
}
\description{
\code{gg_rescale} rescales the given ggplot2 so that the axes follow the
given ratio (horizontally). If a plot is not specified, the last plot
displayed is used.

\code{gg_rescale_golden} rescales a ggplot2 plot so that axes
follow golden ratio

\code{gg_rescale_square} rescales a ggplot2 plot so that its
axes are square
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1)
gg_rescale(ratio = 1.67)

}

