% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_supports_theme_attribute.R
\name{gg_supports_theme_attribute}
\alias{gg_supports_theme_attribute}
\title{Determine whether installed ggplot2 supports a given theme attribute}
\usage{
gg_supports_theme_attribute(x)
}
\arguments{
\item{x}{The name of an attribute (e.g., "plot.caption")}
}
\value{
A logical value indicating whether the given attribute is supported
(\code{TRUE}) or not (\code{FALSE}).
}
\description{
Determine whether installed ggplot2 supports a given theme attribute
}
\examples{
gg_supports_theme_attribute("plot.caption")

}

