% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_golden.R
\name{ggsave_golden}
\alias{ggsave_golden}
\title{Save a plot with proportions equal to the golden ratio}
\usage{
ggsave_golden(filename, plot = last_plot(), ...)
}
\arguments{
\item{filename}{The name of the file to be written to}

\item{plot}{The plot object to save. Defaults to last plot displayed.}

\item{...}{Additional parameters to be passed to \code{\link{ggsave}}}

\item{width}{The width of the resulting image (default: 8)}

\item{height}{The height of the resulting image (default: 4.94)}
}
\description{
\code{ggsave_golden} is a wrapper for \code{\link{ggsave}} that saves a
ggplot to a file with proportions equal to the golden ratio (wide).
}

