% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ratio_golden}
\alias{ratio_golden}
\title{Calculate the aspect ratio of data that would follow the golden ratio}
\usage{
ratio_golden(xvals, yvals, orientation = "horizontal")
}
\arguments{
\item{xvals}{Vector containing the values plotted along the X axis}

\item{yvals}{Vector containing the values plotted along the Y axis}

\item{orientation}{Whether the golden ratio should be "horizontal" or
"vertical" (default: horizontal)}
}
\value{
A numeric value containing the aspect ratio that would produce a plot
with proportions that follow the golden ratio
}
\description{
\code{ratio_golden} returns the aspect ratio (\code{y/x}) that would produce
a plot where the space that covers one unit along the x axis would correspond
to 1.618034 units along the y axis.
}
\examples{
ratio <- ratio_golden(xvals=mtcars$hp, yvals=mtcars$mpg)
}

