% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_ticks}
\alias{remove_ticks}
\alias{remove_ticks_x}
\alias{remove_ticks_y}
\alias{theme_no_ticks}
\alias{theme_no_ticks_x}
\alias{theme_no_ticks_y}
\title{Remove tick marks}
\usage{
remove_ticks(plot = last_plot())

remove_ticks_x(plot = last_plot())

remove_ticks_y(plot = last_plot())

theme_no_ticks_x()

theme_no_ticks_y()

theme_no_ticks()
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}
}
\value{
\code{remove_ticks}, \code{remove_ticks_x}, and \code{remove_ticks_y}
return a modified ggplot2 plot object.

\code{theme_no_ticks}, \code{theme_no_ticks_x}, and
\code{theme_no_ticks_y} return settings for theme elements that result in no
axis tick marks being displayed.
}
\description{
\code{remove_ticks} removes tick marks from both the X and Y axes on the
given plot, while \code{remove_ticks_x} and \code{remove_ticks_y} remove tick
marks from the X and Y axes, respectively. When building a plot, including
\code{theme_no_ticks} will modify the current theme so that no tick marks are
shown. Inluding \code{theme_no_ticks_x} or \code{theme_no_ticks_y} will
remove tick marks on the X and Y axes, respectively.
}
\details{
As a general rule, \code{remove_ticks} is used for an existing plot, and
\code{theme_no_ticks} is used while building a plot.
}
\examples{
# Build a plot. Afterwards, remove the X axis tick marks
p_total <- ggplot(data=msleep, aes(x=vore, y=sleep_total)) +
    geom_point()
remove_ticks_x(plot=p_total)

# The same, but with pipes
library(magrittr)
p_total <- ggplot(data=msleep, aes(x=vore, y=sleep_total)) +
    geom_point() \%>\% remove_ticks_x()

# While building a plot, set the theme to not include X axis tick marks
p_total <- ggplot(data=msleep, aes(x=vore, y=sleep_total)) +
    geom_point() +
    hide_ticks_x()
}

