% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_plot.R
\name{rescale_plot}
\alias{rescale_plot}
\title{Rescale a ggplot2 plot so that axes follow a given ratio}
\usage{
rescale_plot(plot = last_plot(), ratio)
}
\arguments{
\item{plot}{A ggplot2 plot object. By default, the last plot displayed is
used.}

\item{ratio}{The aspect ratio to use for the axes. This is independent of
units used in the plot, so the size of the X axis will be ratio times the
total size of the Y axis.}
}
\description{
\code{rescale_plot} rescales the given ggplot2 so that the axes follow the
given ratio (horizontally).
}
\examples{
p <- ggplot(mtcars, aes(x = cyl, y = mpg)) +
    geom_point(shape = 1)
rescale_plot(ratio = 1.67)
}
\seealso{
\code{\link{rescale_golden}}, \code{\link{rescale_square}}
}

