% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rescale_square}
\alias{rescale_square}
\title{Rescale a ggplot2 plot so that the axes have equal proportions}
\usage{
rescale_square(plot = last_plot())
}
\arguments{
\item{plot}{A ggplot2 plot object. By defauly, the last plot displayed is
used.}
}
\description{
\code{rescale_golden} rescales the given ggplot2 so that the axes have equal
proportions
}
\examples{
p <- ggplot(mtcars, aes(x=cyl, y=mpg)) +
    geom_point(shape=1)
rescale_square()
}

