% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rockpaperscissors}
\alias{rockpaperscissors}
\title{Data set containing locations of simulated rock-paper-scissors players}
\format{A data frame with TODO rows and TODO variables:
\describe{
  \item{Time}{The simulation timestep}
  \item{X}{The X coordinate of the cell on the lattice}
  \item{Y}{The Y coordinate of the cell on the lattice}
  \item{Type}{The type of player: Rock, Paper, or Scissors}
}}
\source{
\url{https://github.com/briandconnelly/seeds}
}
\description{
Simulations were run on a 50x50 lattice. Each player played one strategy.
}

