% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Calculate the distance between the minimum and maximum values}
\usage{
span(n)
}
\description{
\code{span} calculates the distance between a vector's minimum and maximum
values. Only finite and non-NA values are considered.
}
\examples{
vals <- sample(1:100, 10)
span(vals)
}

