% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bdc_lattice_population.R
\name{theme_bdc_lattice_population}
\alias{theme_bdc_lattice_population}
\title{Theme for displaying populations structured on a lattice}
\usage{
theme_bdc_lattice_population(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{The base size for all text (default: 10)}

\item{base_family}{The base font family for all text}
}
\value{
A list of ggplot theme parameters
}
\description{
\code{theme_bdc_lattice_population} is a \code{\link{ggplot2}} theme that
produces a plot where many graphical details such as tick marks and axis
labels and text have been removed. This theme is intended to be used to
display simluated populations where individuals are arranged on a lattice.
}
\seealso{
\code{\link{theme}}
}

