% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bdc_paneled.R
\name{theme_bdc_paneled}
\alias{theme_bdc_paneled}
\title{Clean, borderless theme with grey panels and axis ticks}
\usage{
theme_bdc_paneled(base_size = 12, base_family = "", grid.x = FALSE,
  grid.y = TRUE, gridmin.x = FALSE, gridmin.y = FALSE, ticks.x = TRUE,
  ticks.y = TRUE, pmargin = base_size/2)
}
\arguments{
\item{base_size}{The base size for all text (default: 12)}

\item{base_family}{The base font family for all text}

\item{grid.x}{Show grid lines along the X axis (default: FALSE)}

\item{grid.y}{Show grid lines along the Y axis (default: TRUE)}

\item{gridmin.x}{Show minor grid lines along the X axis (default: FALSE)}

\item{gridmin.y}{Show minor grid lines along the Y axis (default: FALSE)}

\item{ticks.x}{Show tick marks along the X axis (default: TRUE)}

\item{ticks.y}{Show tick marks along th Y axis (default: TRUE)}

\item{pmargin}{Margin around the plot (default: 1/2 line)}
}
\value{
A list of ggplot theme parameters
}
\description{
\code{theme_bdc_paneled} is a \code{\link{ggplot2}} theme that produces a plot
with simple grey panels on a white background. Facet labels are displayed as
black text. Grid lines can be enabled using  the \code{grid.x} and
\code{grid.y} parameters. By default, grid lines along the Y axis are shown.
Tick marks can be disabled using the \code{ticks.x} and \code{ticks.y}
parameters.
}
\examples{
library(ggplot2)

p <- ggplot(movies, aes(x=year, y=budget)) + geom_point() + theme_bdc_paneled()
p

}
\seealso{
\code{\link{theme}}
}

