% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bdc_simplefacets.R
\name{theme_bdc_simplefacets}
\alias{theme_bdc_simplefacets}
\title{Clean theme with black box and axis ticks}
\usage{
theme_bdc_simplefacets(base_size = 12, base_family = "", grid.x = FALSE,
  grid.y = FALSE, gridmin.x = grid.x, gridmin.y = grid.y,
  ticks.x = TRUE, ticks.y = TRUE, pmargin = base_size/2)
}
\arguments{
\item{base_size}{The base size for all text (default: 12)}

\item{base_family}{The base font family for all text}

\item{grid.x}{Show grid lines along the X axis (default: FALSE)}

\item{grid.y}{Show grid lines along the Y axis (default: FALSE)}

\item{gridmin.x}{Show minor grid lines along the X axis (default: FALSE)}

\item{gridmin.y}{Show minor grid lines along the Y axis (default: FALSE)}

\item{ticks.x}{Show tick marks along the X axis (default: TRUE)}

\item{ticks.y}{Show tick marks along th Y axis (default: TRUE)}

\item{pmargin}{Margin around the plot (default: 1/2 line)}
}
\value{
A list of ggplot theme parameters
}
\description{
\code{theme_bdc_simplefacets} is a \code{\link{ggplot2}} theme that produces
a plot in a simple black panel on a white background. Facet labels are
displayed as black text on a transparent background. Grid lines can be
enabled using the \code{grid.x} and \code{grid.y} parameters.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x=hp, y=mpg)) + geom_point() + theme_bdc_simplefacets()
p
}
\seealso{
\code{\link{theme}}
}

