% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_whitespace.R
\name{trim_whitespace}
\alias{trim_whitespace}
\title{Remove whitespace from around an image}
\usage{
trim_whitespace(filename, ...)
}
\arguments{
\item{filename}{Image to be cropped}

\item{...}{Optional arguments}
}
\description{
\code{trim_whitespace} crops an image, removing any whitespace from around
it.
}
\note{
This function requires \href{http://pdfcrop.sourceforge.net}{PDFCrop}
for PDFs and \href{http://www.imagemagick.org/script/index.php}{ImageMagick}
for other formats.
}
\examples{
\dontrun{
trim_whitespace("my_figure.pdf")
}
}

