% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coord_square}
\alias{coord_square}
\title{Set a square aspect ratio}
\usage{
coord_square(xvals, yvals)
}
\arguments{
\item{xvals}{Vector containing the values plotted along the X axis}

\item{yvals}{Vector containing the values plotted along the Y axis}
}
\description{
\code{coord_square} sets the aspect ratio of the axes in the plot to 1:1 (a
square)
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x=hp, y=mpg)) + geom_point() +
         coord_square(xvals=mtcars$hp, yvals=mtcars$mpg)
p
}

