% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_bdc_grey}
\alias{theme_bdc_gray}
\alias{theme_bdc_grey}
\title{Clean theme with grey box and axis ticks}
\usage{
theme_bdc_grey(base_size = 12, base_family = "", grid.x = FALSE,
  grid.y = FALSE, ticks.x = TRUE, ticks.y = TRUE)
}
\arguments{
\item{base_size}{The base size for all text (default: 12)}

\item{base_family}{The base font family for all text}

\item{grid.x}{Show grid lines along the X axis (default: FALSE)}

\item{grid.y}{Show grid lines along the Y axis (default: FALSE)}

\item{ticks.x}{Show tick marks along the X axis (default: TRUE)}

\item{ticks.y}{Show tick marks along th Y axis (default: TRUE)}
}
\value{
A list of ggplot theme parameters
}
\description{
\code{theme_bdc_grey} is a \code{\link{ggplot2}} theme that produces a plot
in a simple grey panel on a white background. Facet labels are displayed as
black text on a grey background. Grid lines can be enabled using
the \code{grid.x} and \code{grid.y} parameters. Tick marks can be disabled
using the \code{ticks.x} and \code{ticks.y} parameters.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x=hp, y=mpg)) + geom_point() + theme_bdc_grey()
p
}
\seealso{
\code{\link{theme}}
}

